+++ config/routes.rb
@@ -44,8 +44,11 @@ Rails.application.routes.draw do
     end
   end
   resources :groups
-  resources :resources, only: :index
-  resources :access_policies, only: [:new, :create, :destroy]
+  resources :resources, only: :index do
+    scope module: :resources do
+      resources :access_policies, only: [:create, :destroy]
+    end
+  end
   resources :computations, only: [:show, :create]
 
   # Help